core/standards/scripts/scheduler2/location_hash/501.html

<!doctype html>
<title>Location hash - document.write location.hash - check in separate code block</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});
</script>
<script>
t.step(function() {
    document.write( "<script>location.hash = '#foo';<\/script>" );
});
</script>
<script>
t.step(function() {
    assert_equals( location.hash, '#foo' );
    assert_equals( history.length, histlength );
});
t.done();
</script>
<div id="log"></div>
